[Core][Mailbox] Fix a potential problem with poll
When users are interleaving the use of two versions of list_execute
(i.e., the one with in/out channels and the one without them),
some communications through the channels may seem to be be 'missing'.
However, actually they are not---it's due to flushing empty
communications. This fix prevents such situation.